Updated message parsing and logging
Instead of url escaping and using an unholy regex to parse fishing
success messages, the parsing is now done with a slightly more readable
finite-state-machine.

The result is also logged as a somewhat more human-readable string, and
unknown log lines are no longer printed. I realized that all of those
messages show up in the ACT window and so I thought they should at least
be informative there. Structured data is available for future use
though.